[FF-1] Update `loadApps` method to load multiple apps instead of one static app
This is no longer a sample method. The `loadApps` method relies on data from the view controller.
It is now the view controller that determines whether the data it passes through is real or a sample.

Likewise it is also the Network manager that takes the app id and acts on it.